Admin Add/Edit/Delete and Reordering of Pages
AdminPagesController.java -- changed to autowire the PageRepository bean rather than initalizing the data in the contructor.

GetMapping and Post Mapping added for Add Page and Edit Page. Delete Page only needed Get Mapping. Delete page has javascript (app.js) to confirm deletion.

When adding page, validation takes place. Title and Content are required. javax.validation is used to check fields.

Since our PageRepository extends JpaRepository, we don't have to write custom queries. Instead, we can use keywords in our method definition to allow the API to distinguish what is needed. Link to info for queries is provided in PageRepository.java.

For sorting, two methods are added to style.css. This allows for admin to reorder pages, add some border for design.

Error handling on context and title fields are also present in the add.html and edit.html. This will be sent to the .java file, when deciding whether there are errors on the fields.

Jquery script is added inside index.html to allow for reordering.