Admin Categories Created
Allow for Adding, editing, and deleting categories on the admin side of the web application.

This follows the same code as the Pages for add, edit, and delete.

Navigation was modified to include the Categories section for the admin.

Database Table Categories was created. Category.java reflects the database table. CategoryRepository.java reflects the methods that allow for communication to the database table by extending the JpaRepository class.