Finished basic requirements for website
User may now use all available functionality for the web application. Admin can also have full functionality over the website. The website still needs to be RESTful. This will be implemented in a later commit. Security was finished in this commit. This allows for login for users and admin. Admin has access to certain pages that a user does not. Users can register now. Database tables have been created to allow for accounts and password encryption is also used for security purposes.